Move resource-based priority functions to their Score plugins
