+++ pkg/scheduler/algorithm/priorities/BUILD
@@ -9,32 +9,22 @@ load(
 go_library(
     name = "go_default_library",
     srcs = [
-        "balanced_resource_allocation.go",
-        "least_requested.go",
         "metadata.go",
-        "most_requested.go",
         "priorities.go",
         "reduce.go",
-        "requested_to_capacity_ratio.go",
-        "resource_allocation.go",
         "selector_spreading.go",
         "test_util.go",
         "types.go",
     ],
     importpath = "k8s.io/kubernetes/pkg/scheduler/algorithm/priorities",
     deps = [
-        "//pkg/apis/core/v1/helper:go_default_library",
-        "//pkg/features:go_default_library",
-        "//pkg/scheduler/algorithm/priorities/util:go_default_library",
         "//pkg/scheduler/framework/v1alpha1:go_default_library",
         "//pkg/scheduler/listers:go_default_library",
         "//pkg/scheduler/nodeinfo:go_default_library",
         "//pkg/util/node:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
-        "//staging/src/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/labels:go_default_library",
-        "//staging/src/k8s.io/apiserver/pkg/util/feature:go_default_library",
         "//staging/src/k8s.io/client-go/listers/apps/v1:go_default_library",
         "//staging/src/k8s.io/client-go/listers/core/v1:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
@@ -44,18 +34,13 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "balanced_resource_allocation_test.go",
-        "least_requested_test.go",
         "metadata_test.go",
-        "most_requested_test.go",
-        "requested_to_capacity_ratio_test.go",
         "selector_spreading_test.go",
         "spreading_perf_test.go",
         "types_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
-        "//pkg/features:go_default_library",
         "//pkg/scheduler/algorithm:go_default_library",
         "//pkg/scheduler/algorithm/priorities/util:go_default_library",
         "//pkg/scheduler/framework/v1alpha1:go_default_library",
@@ -68,12 +53,9 @@ go_test(
         "//staging/src/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/labels:go_default_library",
-        "//staging/src/k8s.io/apiserver/pkg/util/feature:go_default_library",
         "//staging/src/k8s.io/client-go/informers:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes/fake:go_default_library",
-        "//staging/src/k8s.io/component-base/featuregate/testing:go_default_library",
         "//vendor/github.com/google/go-cmp/cmp:go_default_library",
-        "//vendor/github.com/stretchr/testify/assert:go_default_library",
     ],
 )
