--- pkg/scheduler/algorithm/priorities/least_requested_test.go
@@ -1,266 +1,0 @@
-/*
-Copyright 2016 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package priorities
-
-import (
-	"reflect"
-	"testing"
-
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	framework "k8s.io/kubernetes/pkg/scheduler/framework/v1alpha1"
-	nodeinfosnapshot "k8s.io/kubernetes/pkg/scheduler/nodeinfo/snapshot"
-)
-
-func TestLeastRequested(t *testing.T) {
-	labels1 := map[string]string{
-		"foo": "bar",
-		"baz": "blah",
-	}
-	labels2 := map[string]string{
-		"bar": "foo",
-		"baz": "blah",
-	}
-	machine1Spec := v1.PodSpec{
-		NodeName: "machine1",
-	}
-	machine2Spec := v1.PodSpec{
-		NodeName: "machine2",
-	}
-	noResources := v1.PodSpec{
-		Containers: []v1.Container{},
-	}
-	cpuOnly := v1.PodSpec{
-		NodeName: "machine1",
-		Containers: []v1.Container{
-			{
-				Resources: v1.ResourceRequirements{
-					Requests: v1.ResourceList{
-						v1.ResourceCPU:    resource.MustParse("1000m"),
-						v1.ResourceMemory: resource.MustParse("0"),
-					},
-				},
-			},
-			{
-				Resources: v1.ResourceRequirements{
-					Requests: v1.ResourceList{
-						v1.ResourceCPU:    resource.MustParse("2000m"),
-						v1.ResourceMemory: resource.MustParse("0"),
-					},
-				},
-			},
-		},
-	}
-	cpuOnly2 := cpuOnly
-	cpuOnly2.NodeName = "machine2"
-	cpuAndMemory := v1.PodSpec{
-		NodeName: "machine2",
-		Containers: []v1.Container{
-			{
-				Resources: v1.ResourceRequirements{
-					Requests: v1.ResourceList{
-						v1.ResourceCPU:    resource.MustParse("1000m"),
-						v1.ResourceMemory: resource.MustParse("2000"),
-					},
-				},
-			},
-			{
-				Resources: v1.ResourceRequirements{
-					Requests: v1.ResourceList{
-						v1.ResourceCPU:    resource.MustParse("2000m"),
-						v1.ResourceMemory: resource.MustParse("3000"),
-					},
-				},
-			},
-		},
-	}
-	tests := []struct {
-		pod          *v1.Pod
-		pods         []*v1.Pod
-		nodes        []*v1.Node
-		expectedList framework.NodeScoreList
-		name         string
-	}{
-		{
-			/*
-				Node1 scores (remaining resources) on 0-10 scale
-				CPU Score: ((4000 - 0) *100) / 4000 = 100
-				Memory Score: ((10000 - 0) *100) / 10000 = 100
-				Node1 Score: (100 + 100) / 2 = 100
-
-				Node2 scores (remaining resources) on 0-10 scale
-				CPU Score: ((4000 - 0) *100) / 4000 = 100
-				Memory Score: ((10000 - 0) *10) / 10000 = 100
-				Node2 Score: (100 + 100) / 2 = 100
-			*/
-			pod:          &v1.Pod{Spec: noResources},
-			nodes:        []*v1.Node{makeNode("machine1", 4000, 10000), makeNode("machine2", 4000, 10000)},
-			expectedList: []framework.NodeScore{{Name: "machine1", Score: framework.MaxNodeScore}, {Name: "machine2", Score: framework.MaxNodeScore}},
-			name:         "nothing scheduled, nothing requested",
-		},
-		{
-			/*
-				Node1 scores on 0-10 scale
-				CPU Score: ((4000 - 3000) *100) / 4000 = 25
-				Memory Score: ((10000 - 5000) *100) / 10000 = 50
-				Node1 Score: (25 + 50) / 2 = 37
-
-				Node2 scores on 0-10 scale
-				CPU Score: ((6000 - 3000) *100) / 6000 = 50
-				Memory Score: ((10000 - 5000) *100) / 10000 = 50
-				Node2 Score: (50 + 50) / 2 = 50
-			*/
-			pod:          &v1.Pod{Spec: cpuAndMemory},
-			nodes:        []*v1.Node{makeNode("machine1", 4000, 10000), makeNode("machine2", 6000, 10000)},
-			expectedList: []framework.NodeScore{{Name: "machine1", Score: 37}, {Name: "machine2", Score: 50}},
-			name:         "nothing scheduled, resources requested, differently sized machines",
-		},
-		{
-			/*
-				Node1 scores on 0-10 scale
-				CPU Score: ((4000 - 0) *100) / 4000 = 100
-				Memory Score: ((10000 - 0) *100) / 10000 = 100
-				Node1 Score: (100 + 100) / 2 = 100
-
-				Node2 scores on 0-10 scale
-				CPU Score: ((4000 - 0) *100) / 4000 = 100
-				Memory Score: ((10000 - 0) *100) / 10000 = 100
-				Node2 Score: (100 + 100) / 2 = 100
-			*/
-			pod:          &v1.Pod{Spec: noResources},
-			nodes:        []*v1.Node{makeNode("machine1", 4000, 10000), makeNode("machine2", 4000, 10000)},
-			expectedList: []framework.NodeScore{{Name: "machine1", Score: framework.MaxNodeScore}, {Name: "machine2", Score: framework.MaxNodeScore}},
-			name:         "no resources requested, pods scheduled",
-			pods: []*v1.Pod{
-				{Spec: machine1Spec, ObjectMeta: metav1.ObjectMeta{Labels: labels2}},
-				{Spec: machine1Spec, ObjectMeta: metav1.ObjectMeta{Labels: labels1}},
-				{Spec: machine2Spec, ObjectMeta: metav1.ObjectMeta{Labels: labels1}},
-				{Spec: machine2Spec, ObjectMeta: metav1.ObjectMeta{Labels: labels1}},
-			},
-		},
-		{
-			/*
-				Node1 scores on 0-10 scale
-				CPU Score: ((10000 - 6000) *100) / 10000 = 40
-				Memory Score: ((20000 - 0) *100) / 20000 = 100
-				Node1 Score: (40 + 100) / 2 = 70
-
-				Node2 scores on 0-10 scale
-				CPU Score: ((10000 - 6000) *100) / 10000 = 40
-				Memory Score: ((20000 - 5000) *100) / 20000 = 75
-				Node2 Score: (40 + 75) / 2 = 57
-			*/
-			pod:          &v1.Pod{Spec: noResources},
-			nodes:        []*v1.Node{makeNode("machine1", 10000, 20000), makeNode("machine2", 10000, 20000)},
-			expectedList: []framework.NodeScore{{Name: "machine1", Score: 70}, {Name: "machine2", Score: 57}},
-			name:         "no resources requested, pods scheduled with resources",
-			pods: []*v1.Pod{
-				{Spec: cpuOnly, ObjectMeta: metav1.ObjectMeta{Labels: labels2}},
-				{Spec: cpuOnly, ObjectMeta: metav1.ObjectMeta{Labels: labels1}},
-				{Spec: cpuOnly2, ObjectMeta: metav1.ObjectMeta{Labels: labels1}},
-				{Spec: cpuAndMemory, ObjectMeta: metav1.ObjectMeta{Labels: labels1}},
-			},
-		},
-		{
-			/*
-				Node1 scores on 0-10 scale
-				CPU Score: ((10000 - 6000) *10) / 10000 = 40
-				Memory Score: ((20000 - 5000) *10) / 20000 = 75
-				Node1 Score: (40 + 75) / 2 = 57
-
-				Node2 scores on 0-10 scale
-				CPU Score: ((10000 - 6000) *100) / 10000 = 40
-				Memory Score: ((20000 - 10000) *100) / 20000 = 50
-				Node2 Score: (40 + 50) / 2 = 45
-			*/
-			pod:          &v1.Pod{Spec: cpuAndMemory},
-			nodes:        []*v1.Node{makeNode("machine1", 10000, 20000), makeNode("machine2", 10000, 20000)},
-			expectedList: []framework.NodeScore{{Name: "machine1", Score: 57}, {Name: "machine2", Score: 45}},
-			name:         "resources requested, pods scheduled with resources",
-			pods: []*v1.Pod{
-				{Spec: cpuOnly},
-				{Spec: cpuAndMemory},
-			},
-		},
-		{
-			/*
-				Node1 scores on 0-10 scale
-				CPU Score: ((10000 - 6000) *100) / 10000 = 40
-				Memory Score: ((20000 - 5000) *100) / 20000 = 75
-				Node1 Score: (40 + 75) / 2 = 57
-
-				Node2 scores on 0-10 scale
-				CPU Score: ((10000 - 6000) *100) / 10000 = 40
-				Memory Score: ((50000 - 10000) *100) / 50000 = 80
-				Node2 Score: (40 + 80) / 2 = 60
-			*/
-			pod:          &v1.Pod{Spec: cpuAndMemory},
-			nodes:        []*v1.Node{makeNode("machine1", 10000, 20000), makeNode("machine2", 10000, 50000)},
-			expectedList: []framework.NodeScore{{Name: "machine1", Score: 57}, {Name: "machine2", Score: 60}},
-			name:         "resources requested, pods scheduled with resources, differently sized machines",
-			pods: []*v1.Pod{
-				{Spec: cpuOnly},
-				{Spec: cpuAndMemory},
-			},
-		},
-		{
-			/*
-				Node1 scores on 0-10 scale
-				CPU Score: ((4000 - 6000) *100) / 4000 = 0
-				Memory Score: ((10000 - 0) *100) / 10000 = 100
-				Node1 Score: (0 + 100) / 2 = 50
-
-				Node2 scores on 0-10 scale
-				CPU Score: ((4000 - 6000) *100) / 4000 = 0
-				Memory Score: ((10000 - 5000) *100) / 10000 = 50
-				Node2 Score: (0 + 50) / 2 = 25
-			*/
-			pod:          &v1.Pod{Spec: cpuOnly},
-			nodes:        []*v1.Node{makeNode("machine1", 4000, 10000), makeNode("machine2", 4000, 10000)},
-			expectedList: []framework.NodeScore{{Name: "machine1", Score: 50}, {Name: "machine2", Score: 25}},
-			name:         "requested resources exceed node capacity",
-			pods: []*v1.Pod{
-				{Spec: cpuOnly},
-				{Spec: cpuAndMemory},
-			},
-		},
-		{
-			pod:          &v1.Pod{Spec: noResources},
-			nodes:        []*v1.Node{makeNode("machine1", 0, 0), makeNode("machine2", 0, 0)},
-			expectedList: []framework.NodeScore{{Name: "machine1", Score: 0}, {Name: "machine2", Score: 0}},
-			name:         "zero node resources, pods scheduled with resources",
-			pods: []*v1.Pod{
-				{Spec: cpuOnly},
-				{Spec: cpuAndMemory},
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			snapshot := nodeinfosnapshot.NewSnapshot(nodeinfosnapshot.CreateNodeInfoMap(test.pods, test.nodes))
-			list, err := runMapReducePriority(LeastRequestedPriorityMap, nil, nil, test.pod, snapshot, test.nodes)
-			if err != nil {
-				t.Errorf("unexpected error: %v", err)
-			}
-			if !reflect.DeepEqual(test.expectedList, list) {
-				t.Errorf("expected %#v, got %#v", test.expectedList, list)
-			}
-		})
-	}
-}
